As part of the LLVM3.0 migration we are updating our machines to gnuwin32, which is a set of unix tools used to run tests by llvm.
As part of the migration from Cygwin to gnuwin32, 7 tests started passing again. This is good, however the transition to gnuwin will be gradual and we can't have tests failing during the process. 

We remove these tests (from the 2.8 branch only) in order to enable our migration.

Together with Vlad.