Fixed CQ ticket CSSD100012114: Unnecessary Barrier call leads to incorrect output.
The bug was in WIRelatedValue pass when it returns the answer "not WI related" for Alloca instructions assuming that it will be handled manually later in the barrier passes.
However, it did not take into account that this decision will affect other instructions like Gep and Load that use this Alloca and need to be assume "WI relaled" themselves.
Fixed by changing WIRelatedValue pass to return "WI related" for Alloca Instruction.

Also, added few LIT tests that covers this case.

Reviewed by Ran & Sion