CSSResolved CQ ticket CQ100012223: check_barrier is failing on Linux...
Removed the "opt.exe" flag "-print-module" from barrier LIT tests, which was in use for debugging the tests and as part of it.
This flag makes "opt" prints the module to the standard error stream, and it seems that any print to the error stream in Linux makes the LIT test fail.

(Passed pre-commit)

Reviewed by Sion