Revert 31793 which re-applied r31512 which was reverted in 31578 per request from the PDT performance forum, until further investigation of the performance regressions in intel_median.
Original message:

Enable aggressive unrolling early in the O3 optimization for scalar code. This allows SRA to do a much better job because complete loops are eliminated. These loops often access small local arrays which are better represented as multiple scalars. 

Reapply r31512 which was reverted in 31578. It improved a number of WOLF and SHOC workloads. We do have a regression in intel_median using the auto_detect mode, but see a 1.4x boost on scalar mode.