Fix for CSSD100011811 - problem in casting metadata elements
Description:

The original bug was opened a while ago, and we do not have a test that reproduces it. Nevertheless, we can adjust the code to avoid this issue again and to reflect the current behavior of the metadata.
There was an exception due to the fact that one of the metadata nodes of "opencl.wrappers" was not really a function and the result of the llvm::dyn_cast of that node was not checked.
Now, the only metadata elements we should encounter for "opencl.kernels", "opencl.wrappers" are llvm::Function, so we can use llvm::cast instead of llvm::dyn_cast. In case of bad cast, it will cause an assertion failure.

ClearQuest: CSSD100011811

Reviewer: Guy, Eli (discovered the bug, suggested the solution)

ReviewBoard: http://cvcc-ubu-01.iil.intel.com/cr/r/1005/

Pre commit: http://cvcc-fc14-01.iil.intel.com:8080/job/PersonalCompilerReport/771/

Performance: No performance impact is expected