This commit contains two fixes that fix many tests in IntegerOps
The first one sets the higher bytes to zero before a store with down convert to a short or a char.
The reason is that the down convert store does saturation and does not truncate the higher bytes.
Therefore, in case of an overflow the highest possible value is being written instead of the value represented by the lower bytes.
The second one adds specific extention in case of a promoted type for udiv, sdiv, urem, and srem.
The reason is that for these operations a specific sign/zero extend should be done and not any extend.

code reviews: http://cvcc-ubu-01.iil.intel.com/cr/r/1091/ and http://cvcc-ubu-01.iil.intel.com/cr/r/1120/

testing: mic's precommit and the general precommit.