Fixed DAGCombiner issue for performing zext upconverting load when loading unsigned int vectors.
When a uint8 or uint16 vector is loaded, the elements are extended with anyext and then the vector is ANDed with a masking vector to zero the upper bits.

DAGCombiner's visitAND attempted to fix it by identifying ANDs with a splatted-constant vectors that immediately follow a load and replacing it with a zext-ing load.

However there was a bug there which prevented the splatted constant vector from being identified.

The bug is now fixed, which also enables MIC's upconverting loads.