This commit contains some utilities for development mainly on Apple environment. None of this changes happens on standard compilation, it is only meant to be used when developing transformations or debugging.
PrintfAdder - Utility to add printf calls over arbitrary llvm instructions which is useful when we don't have the cl source, or would like to print masks (which are not in the source), or values in masked lanes.
IRInjection - Injects IR from a file, but keeps all pointer to functions, global variables valid.
Namer - changed it to set names to both instruction and basic blocks according to the instruction opcode and type of the instructions. This makes it easier to compare difference in IR.

Pre -commit passed (except for failure in debugger test type that happened also in post-commit)