Fix for handling of #praga FP_CONTRACT off
The current handling of this pragma is incorrect since the built-ins module is examined instead of the module containing the kernel.
The best solution would be initializing the execution engine every time we compile a kernel with the mic_fma feature on or off depending on the existence of the pragma, like what happens for the CPU device. However, we do not support this mode on MIC (CSSD100014184), so a second best solution is passing the flag using the PiggyBackAnalysis pass.

CR: http://cvcc-ubu-01/cr/r/1253/