Description:
This commit contains Alias analysis for openCL that does two things:
1. Reports pointers to different address spaces as no alias.
2. Overcomes the restrict information lost caused by the wrapper and the inlining.

In order to perform the second bullet, the wrapper was augmented to mark the actual parameters for the previously restricted formal arguments as restricted.
Later on, the new alias analysis utilizes this information in order to increase its accuracy.

In addition, the memory disambiguation for PCG was implemented and utilizes the new alias analysis.

This version is only for MIC, I will check why non-immutable alias analysis are not returned using getAnalysis and add a fix for the CPU.

Performance:

The benchmarks in which we saw performance improvment of more than 10% and their corresponding improvement percentages are:

Reduction_DP 20%
Reduction 10%
SGEMM-N 12%

Code Review: http://cvcc-ubu-01.iil.intel.com/cr/r/1518/

Testing: Precommit