Conformance and CPU update. Merge rev 50337, 50993 from beta branch.
The fix to the conformance is committed on behalf of VPG.
The fix to the CPU part is committed on behalf of me.

The changes changes rounding mode to "nearest even" for index for 1D array and 2D array images.
The changes remove a normalization of the array index for 1D array and 2D array images.

Affected projects
=================
* All projects build

Relevant Bugs
=============
* CSSD100014835

Reviewers
=========
* Bartosz Sochacki(conformance part)

Tests
=====
* Jenkins Pre-Commit.
* All Images tests.

Pre-commit: passed.