This commit contains the changes done at revision 49017 (which reverted later on revision 49262):
Reduced implicit parameters passed to kernel wrapper.
Current implementation requires Backend to pass implicit parameters to the kernel wrapper as part of the structure of parameters.
Two of these implicit parameters are:
  1. The implicit local buffer that is being allocated by the runtime on a request from the backend.
  2. Place holder for current work item, that is used sometimes by the barrier implementation.
Both parameters can be allocated by the wrapper pass, instead of load them from the structure of parameters.
Also reduced the size of the structure accordingly.