Fix for CSSD100015140: Packetizer incorrectly preserves alignment of zero.
The bug:
For LLVM loads and stores an alignment value of 0 or an omitted align argument means that the operation has the abi alignment for the target. The Packetizer copies the alignment of the pre-packetized instruction to the alignment of the packetized instruction, but this breaks correctness for the case of an alignment a value of 0 or an omitted align argument. This bug is not easy to reproduce since Clang usually places explicit alignment on loads/stores.

The fix:
When packetizing loads/stores make sure we always specify alignment explicitly.

CR: http://cvcc-ubu-01/cr/r/1776/