Fix to CSSD100014903 and CSSD100015166
The way we treat local memory today is as follows:
First, Clang translates local memory to global memory. 
Then, the local memory is being transferred as an argument to the kernel and all the references to the global memory are updated to point to the new argument.

The problem is that no one removes the global memory that was added by clang and we allocate space for this memory in the jit.
Victor in his stencil example generated huge local memory which exceeds the size we allocate for the jit.

What I did is adding a GlobalOptimizerPass after the local memory translation to make sure that the global memory allocated by clang is removed.

Code review: http://cvcc-ubu-01.iil.intel.com/cr/r/1777/