Until now, we predicated the entire kernel when it had a divergent branch. This commit enables partial predication and does not predicate the unneeded parts. The idea is that uniform branches without incoming masks won't be predicated and the control flow will remain as it is. The motivation for this feature is to: (a) remove redundant mask manipulation, (b) refrain from executed unneeded instructions, and (c) refrain from paying the penalty of masked operations that can cause in some cases to either scalarization or to pay the penalty of operations as gather/scatter.
In order to do the above we had to do the following: (a) augment the work item analysis to propagate control flow divergence into the data, 
(b) modify the predication to not predicate uniform branches without incoming masks, and (c) modify the specializer not to add bypasses for the branches from (b).

Overall, on CPU, this commit provides a 3% speedup while running Cyberlink on Haswell and 5.5% speedup while running Cyberlink on Sandybridge.
Even though the geomean of every benchmark does not show degradation, we experienced degradation in some of the examples.
Some of these degradations caused by a difference heuristic decision while others need to have further investigation.
Some of the *Vert* samples from CyberLink showed degradation and needs to be fixed.

On MIC, we show performance improvement of 1.24x on SHOC's Scan DP, 2.66x on SHOC's Sort, 3.66x on convertible bonds, 1.43x on an optimized version of BUDE, 
and 2.45x on the original BUDE. The only degradation we experienced was of 12% on SHOC's Scan SP.

Code review: http://cvcc-ubu-01.iil.intel.com/cr/r/2172/ and http://cvcc-ubu-01.iil.intel.com/cr/r/2051/

Testing: precommit