Fix of fp-contact default in CPU BE
The default was enforced to be "disabled" in metadata, while in compiler fma generation was defaulted to "enabled". Both mistakes created the situation when unless CLANG set "enabled" - no fma was produced. The bug was exposed when CLANG was modified.
The current protocol is as follows:
1. CPU BE doesn't produce fp-contract metadata by default.
2. CLANG produces fp-contract metadata as "enabled" by default (in accordance with OCL & SPIR specs).
3. CPU BE uses this metadat for fma code generation.
4. If metadata are missed, then the default is "disabled" (in accordance with OCL spec).

CR: http://cvcc-ubu-01.iil.intel.com/cr/r/2331/
Pre-commit: http://cvcc-fc14-01.iil.intel.com:8080/job/PersonalCompilerReport/6/