Fix of compiler crash during vectorization. Root-case:Linearizer logic lacked PHI node fix-up for the following case: a source block in a loop has 2 outgoing edges, which  are not the loop headers. In that case there was no fixup of PHI nodes of the block which as successive to it in the course of linearization, and that broke coherency between PHI-node and post-linearization control flow. Fix:The fix is to add PHI node fixup of the successive block. Test:The bug was exposed by an external testcase, which is added as well (as a LIT test).
The bug was dormant since Vectorizer introduction, because prior to WI analysis (which leaves CFG as is for uniform cases) entire function was linearized. As the problematic case can happen deep down in the flow (as CFG compexity increases by then), then total linearization (prior to that point) prevented the bug exposure - it simplified CF as long as it traversed the function's BBs. When WIA was introduced, a complex CFG could reach the point where linearization is relevant. This is exactly the case of LIT test.
Also, the LIT test is somehow unusual for OCL - because it uses nested loops instead of implicit WI loop. That was another trigger for the bug exposure.

CR: http://cvcc-ubu-01.iil.intel.com/cr/r/2435/
pre-commit:http://cvcc-fc14-01.iil.intel.com:8080/job/PersonalCompilerReport/116/