The current bypass support has a few problems: 1. The RegionInfo analysis that is used for bypasses is obsolete and needs to be replaced 2. There are currently no bypasses on regions build by a single basic block 3. Bypasses are not only generated after branches, which cause generation of redundant bypasses and consequently to runtime overhead 4. Regions generated by RegionInfo are too small which reduces the number and effectiveness of bypasses
This commit should fix all the above by adding bypasses only after branches and calculating the exit for a region starting at a branch successor s, by taking the last post-dominator of s that is also dominated by s.
In case that the first post-dominator of s does not dominated by s, s is marked as a singleton region.
In addition, this commit adds bypasses for regions with two outgoing edges.

Currently the results we got are 3%-4% on SHOC, 1.5% - 2.2% on CyberLink, 0.2 - 3.8% on WOLF, 1.8% - 3.6% on RightWare (depend on the architecture).

Code Review: Amjad and Zvi

Testing: precommit