In case a branch condition becomes random during branch divergent propagation its divergency should continue propagating. This happens in rare cases, as in do-while, where the exit basic block of the loop has a branch that its divergency is depend on the loop divergency.
This does not happened in regular loops because in this case, the exit node will contain a PHI that will be marked as RANDOM and will pollute its dependent branch in the regular flow.

This fix makes sure the propagation continues in such a case.

This bug popped out during Mohammed's vectorizer port to RenderScript

Code review: http://cvcc-ubu-01.iil.intel.com/cr/r/2545/

Testing: precommit