This commit fixes CSSD100016682.
When doing a partial predication (predicate only the divergent branches of a kernel) we need to guarantee that the linearizer will not accidentally add an unconditional branch from a predicated to a non-predicated node.

This commit forces the linearizer to put basic blocks from the same predicated regions one after the other.

Note that we add constraints only for the largest regions and not for their sub-regions. 
This is done from two reasons:
1. It is enough to do so :)
2. The sub-regions might have common basic blocks and therefore, adding constraints for each one of them may cause to a conflict at the linearizer. 

Adding a constraints for the largest region solves this second problem because WIA guarantees (by polluting uniform branches if needed) that if two predicated regions have common BBs then the minimal region that includes both of then will also be predicated.
This is done in order to guarantee the correctness of predication and helps us in here.

Also note that for predicated regions appear one after the other we also have to add one constraint for all these regions together.

Code review: http://cvcc-ubu-01.iil.intel.com/cr/r/2795/

Testing: precommit