This commit aims to fix regressions in the original version of SG++ (CSSD100016445).
Today the bypass heuristic decides to put a bypass over a single basic block if it contains a function call.
This creates a redundant bypass in case that the called function has a very small computation cost (and lateron will be inlined).

I added a map that notifies the heuristics whether a call has a small computation cost. 

The map can be augmented with additional builtins (a tuning work can be done after the gold+1 release).

Code review: http://cvcc-ubu-01.iil.intel.com/cr/r/2842/

Testing: precommit and performance