Changes in the Barrier pass
[[1]]
Problem: Using std::map::iterator and std::set::iterator to iterate over data structures which have keys of pointer type causes a differenece in resulting IR and generated code from one run to another. This is problematic since it may cause variations in performance measurements and makes it more difficult to pinpoint a commit which caused a regression, and is not convenient in general during development.
Solution: use llvm::VectorSet and llvm::VectorMap which have the following pros:
1. They guarantee iteration in same order as insertion order.
2. Iteration over elements is more efficient, since a vector is being traveresed instead of a tree.
and the cons are:
1. They require twice as much space to store elements: tree + vector
2. After insertion/deletion of element, the iterators are invalidated. This is a very important attribute that requires special attention when using these data structures. Some modifications to the code were made because of this, mainly instead calling operator[] for checking existence of key, use a const getter such as count()

[[2]]
Remove some redundant dyn_cast<>'s

CR: http://cvcc-ubu-01/cr/r/2907/
Pre-commit: http://cvcc-fc14-01:8080/job/PCompiler_zrackove/18/