Tests: (*) Pre-commit of Profiler on Jenkins     (it runs conformance tests)
Reviewers:
(*) Oren Sarid
(*) Oded Perez
(*) Arik Adler

Review board tickets:
- 2894
- 2922
- 2936
- 2937


======================
API Debugger changes
======================
- Logger: added option to dump messages to log file in real-time
- Logger: added lots of debugging messages on various places in the code
 (to facilitate handling with unexpected errors)

- Trace view: added option to toggle between "function name" (compact view) 
and "function with arguments" in the message column
- Trace view: added option to dump trace output to a CSV file
- Trace view: added "Time' column do indicate when did the API occur (server time)

- All views: shortened titles of view windows from: "OpenCL API Debugger - <view name>"
to: "<view name>" so that it's easier to determine the window name when many windows are opened simultaneously

- Model: fixed a bug in the command-queue model in which (enqueued) commands would not 
enter the queue if the first callback received from them wasn't CL_SUBMITTED.
Now allowing command to start from CL_RUNNING or CL_COMPLETE

- Communicator (sockets): cleaned up receive method, added a blocking method to handle both header and message
  (timeout has been set as well so that we don't wait forever for the message to arrive)

  
  
  
  
========================================
 Profiler & oclServerNotifier changes
========================================
The following changes have been made:
1) All OpenCL 1.2 APIs are traced & monitored (via registered notifiers)
2) Customized the Notifier collection APIs
3) Various small bug fixes
4) Wrapped all GPA code in USE_GPA_TRACING macro to enable
   easy removal of this component if decided not to use it in the future
5) Synced clCreatePerfCountersCommandQueueINTEL API with VPG
   A new parameter "configuration" was added to the function signature.
   Passing configuration=2 preserves the expected behvaiour of the profiler
   before this parameter was introduced.
   
   
   
   
------------------------------
Source code important changes:
------------------------------

=============================
2) dispatch.cpp: wrapped all GPA-related macros/code inside APIs,
=============================

3) notifierservermessages.*
the .pb.h and .pb.cc, are pre-compiled files generated from notifierservermessage.proto.

As for notifierservermessage.proto, I've added a dateTime message to indicate to the client
at which time have the API been processed.
=============================

4) oclInternalFunctions.h
Describes APIs that should not pass through the tracing mechanism.
Since I cannot overload the API calls of OpenCL I had to duplicate the code of the APIs
and remove the code that traces them.
=============================

5) oclServerNotifier.h

See (9), API interface changes.

=============================
6) cl_socket.cpp
- Added a command to get the full string name of the error
(useful when it crashes, I can actually see the error message in the string variable)

=============================
7) oclServerNotifier.cpp
- Changed all OpenCL API calls (used for retrieving info about objects) to use the internal functions
(those that are not traced)



=============================
8) oclNotifier.cpp
- releaseCommandData
unified the following 2 functions:
(*) void NotifierCollection::releaseCommandData(cl_event *event, CommandData* data)
(*) void NotifierCollection::releaseCommandData( cl_event event, CommandData* data )

into:
void NotifierCollection::releaseCommandData(CommandData* data)

one function that handles the release well, and reclaims memory only when it is actually ok.


- enableKernelArgumentInfo
Fixed a bug where the string for the argument to search had a leading space
and it was therefore not found if it was the only argument.



=============================
9) oclNotifier.h

- CommandData struct

Added pEvent and needRelease to track more carefully of the memory objects
needed to be released later.
Added funcName to allow sending the command names of commands that do not appear
in CL_EVENT_COMMAND_TYPE (such as clEnqueueMarkerWithWaitList)


- EventCreate signature change
the notifier client wants to know the name (type) of the enqueue command associated with the event, and some of the commands (such as clEnqueueMarkerWithWaitList) don't return the name of the command, so clGetEventInfo would fail, therefore I'm manually passing the function name.

- ProgramCreate signature change
for some API calls it not necessary to deliver the binary/source, so I've added flags for that (to avoid overhead).



============================
10) dispatch.h
(*) Added member "event" to CallbackParamType in order to release the CL resources properly when we retain a user event (i.e. we wish to decrease reference count of both Intercept & Runtime and maintain the coherency between them).

When we create the user event, we don't use this member, instead we use the RT routines directly (clSetEventCallback and clReleaseEvent).

============================
11) intercept.h & intercept.cpp
(*) Added initialization parameter to createInterceptorEvent()
to keep Intercept coherent with the Runtime reference count of the event.
It happens when user provided us a valid event and we retain it.