Fixed two  predicator correctness bugs in the masks of loops.
Short Description: (Longer description follows at the end)
1) When a loop has a single block exit, but this exit is nested inside a condition, the new loop in-mask was falsely set to represent only the work-items that reached the exit block.
2) When the same block exits more than one loop level, the out-mask represented only the work-items that exited the loop at the last iteration of the second-lowest loop level.

Code Review: Amjad, Victor
http://cvcc-ubu-01.iil.intel.com/cr/r/2938/

Pre-commit:
http://cvcc-fc14-01.iil.intel.com:8080/job/PersonalCompilerReport/998/

Potential Performance Implications:
When the code is vectorized, it is possible that some loops were exited prematurely because of this correctness bug. Thus, an unaviodable regression might occur.

Longer Description:
Bug 1:
When exiting a loop, the in-mask for the loop is updated. There are two different ways to update it.
A "heavy" way: new loop-in-mask = old loop-in-mask & (!(exit-block-in-mask & exit-condition))
A "light" way: new loop-in-mask = (exit-block-in-mask & local-condition)
The heavy way is the "correct" one, and the light way is an optimization that can be used
if exit-block-in-mask == old loop-in-mask. (You are invited to test your De-Morgan capabilities ;))
Before the fix, the code assumed that if there is only one exit block to the loop, then exit-block-in-mask == old loop-in-mask.
However, this is incorrect. (Consider the loop: while (Cond1 || Cond2) {...})
This bug can potentially cause too partial masks including too early loop exits.
The fix checks that the optimization will be used only when we really know that exit-block-in-mask == old loop-in-mask,
including cases in which there are more than one block exiting the loop (which were un-optimized before).

(remarks: exit-condition = !local=condition; 
the expression (exit-block-in-mask & local-condition) is always computed and referred to as localOutMask;
the expression (exit-block-in-mask & exit-condition) is always computed and referred to as who_left_tr)

Bug 2:
When a loop has several exit blocks, out-masks are maintained to reflect which work-item used which exit.
These masks are initialized to false (all-zero) before entering the loop.
The old code initializes these masks in the pre-header of the inner-most loop of the exit-block. This is erroneous in case the same block exits more than one loop - the mask will only reflect the work-items that exited at the last iteration of the second-lowest-level loop.
Instead, these masks should be initialized in the pre-header of the highest-level loop the exit block exits.

(remark:
The case in which the same block exits more than one loop level may at first glance seem unnatural, but this can be the result of compiler actions. Consider the code: While (...) { ... if (...) {...}} Originally it holds only a single loop, but after some compiler work it may become: while-outer -> while-cond <-> while body -> if then -> while-outer.
And the while-cond block could also lead outside the loop, making it a single block exiting both of the loops)