1. The Barrier Buffer is used for saving live values of the current work item before switching to the next work item. This commit changes how this buffer is allocated. Instead of using the memory allocated externally by the Runtime, allocate the memory on the JIT's stack.
2. Similiarly, allocate dynamic local memory (__local arguments to kernel) on the JIT's stack.

Note 1: the interface with the Runtime for memory usage is still needed in order to implement OpenCL queries such as CL_KERNEL_PRIVATE_MEM_SIZE

Note 2: this commit does not make in any changes in the Runtime, so the Runtime will still allocate the buffers mentioned above, even though these allocation will not be consumed. The Runtime team will make the required changes later.

This commit relies on the following commits to the LLVM branch:
76063, 76640, 77070, 77147, 77160+77169, 77773, 77857+77953+77957

CR: http://cvcc-ubu-01/cr/r/3037/
Pre-Commit: http://cvcc-fc14-01:8080/job/PersonalCompilerTests/3601/