WA fix for CSSD100016136: "[Parboil] Parboil/histo killed by OS due to memory leak"
Parboil test killed by OS because of memory explosion in clEnqueueWriteBuffer() when its blocking. When this API is blocking we allocate an internal buffer (to copy user's data) with alignment of 64. For some reason, the allocated space is not freed in the free() call. When the allocation's alignment is changed (to 128 for example) the problem is solved, failed to figure the cause of such behavior so we are going with this WA.

Reviewer: Dima

PreCommit:
http://cvcc-fc14-01.iil.intel.com:8080/job/PRuntime_nmeraey/104/