Optimizer Refactor
With the OpenCL 2.0 compilation mode, we have a fail-check point during the compilation where we decide if we can complete the compilation or fail the compilation.
Reduce the numbers of PassManagers from 3  to 2. One is run before the fail-check point and another is conditionally run after the fail-check point.
Separate code for setting up the passmanagers to separate functions to avoid accidents such as those that exist in the current code where the code for setting up of two different pass managers is interleaved.

Next steps: revisit each pass-manager and shuffle pass order where neccessary.

CR: http://cvcc-ubu-01.iil.intel.com/cr/r/3430/
Pre-commit: http://cvcc-fc14-01.iil.intel.com:8080/job/PersonalCompilerReport/1292/