This commit contains three changes that make the heuristics decide to not vectorize GEHC's kernels, which is the right decision.
The first change is in the weights of transpose function.

The second and third changes were done to let SCEV identify constant trip counts of predicated vectorized loops. Because it cannot do so, the trip count for the predicated vectorized loop used by the heuristics was 32, while for the scalar loop was 818, which made the heuristics make the wrong decision. 

The change I did in the predicator was to leave loop's uniform exit conditions as there are. However, this was not enough because WIA mistakenly marked the induction variable of a loop with a constant number of iterations as random.
To overcome this I changed the WIA that instead of marking each phi node in a partial join as divergent. It marks the phi node as divergent in two cases: the first case is when there exist two paths from different successors of the divergent branch to a different predecessor of the partial join. And the second case is when a definition in the influence region is used in the partial join by a phi node.

Code review: http://cvcc-ubu-01.iil.intel.com/cr/r/3477/

Testing: precommit