Merge rev. 80200 from R3 branch.
"
Message:
Added smartGVN pass to configure LLVM's global value numbering optimization (GVN-PRE) so as not to move loads out of loops that already suffer from high register pressure.

The issue with GEMM performance was exposed by optimization at rev. 79101.
Problem description:
With 79101 we allocate dynamic local memory (__local arguments to kernel) on the JIT's stack. This enables alias analysis and eventually GVN-PRE to hoist some of load instructions from the inner loop of GEMM workload.
Unfortunately, LLVM generates a lot of spills in that case, which kill the performance of GEMM.

Proposed solution:
GVN pass has a parameter that disables this optimization for load instructions.
We add new pass that analyses the function's code and looks for the cases similar to GEMM. Once it finds it - we disable GVN-PRE for loads, otherwise run regular GVN pass.
"

CQ: CSSD100017465
Pre-Commit: http://cvcc-fc14-01.iil.intel.com:8080/job/PersonalCompilerReport/1396
Performance impact: SHOC/GEMM 2x improvement is expected.