Fixing the llvm 'regression' in CloneFunctionInto function.
The problem is that using the CloneFunctionInto in BuiltInFuncImport.cpp causes the function and function declaration attributes to be merged unconditionally causing a problem if both function and function declaration have conflicting attributes. This was the case with FMA builtin where the declaration has ReadOnly and definition has ReadNone attributes - merging those attribute results in conflicting set of attributes.

The patch was to watch for such conflicts after using CloneFunctionInto and fix the conflict if necessary.