master: CSSD100016517 workaround (from releases/2013_hsw_pv_refresh).
This is a merge and redesing of the same workaround from HSW refresh branch.
This workaround is to be used as a base for development of a complete
fix of clGetKernelWorkGroupInfo with  CL_KERNEL_WORK_GROUP_SIZE and
CL_KERNEL_PRIVATE_MEM_SIZE.

Before this fix the query could return max. possible work-group size even
if there was a huge array in the private memory. This caused stack overflow
in Sandra. The workaround is using Barrier pass to calculate an approximate
private memory usage. Most probably this part is to be reverted.

Review:   http://cvcc-ubu-01.iil.intel.com/cr/r/3941/
          aaboud, aharon, omaslov

Pre-Commit: http://compute-ci-jenkins.iil.intel.com:8080/job/PCompilerReport/134/

No noticable performance implications are expected.