Resolved CSSD100019006: [ Intel HPC Samples ] 0.3x-0.5x and lower regressions at 17734 (Trunk)
Previously, WIAnalysis was modified to calculate GEP WI dependency as PTR + UNI = STR, when the GEP type is different than the GEP pointer.
However, there is an exception for that rule, which is SOA Alloca related pointer.
SOA Alloca will be turned into SOA format. Thus, it is allowed to assume: PTR + UNI = PTR