All-ones bypasses optimization. Masked instructions (mostly load/store) takes longer to execute than unmasked instructions even when the mask is full (allones).
The allones bypass is basically a dynamic test, checking whether the mask is allones, and if it is, execute a non-predicated version of the code.

The allones bypasses are implemented on each BasicBlock separately. That is, if a heuristic decides that we want to insert an allones bypass for a block, then we add an 'entry' block that tests whether the mask is allones, then branch to a predicated / or non-predicated version accordingly, and then unite back again. i.e.:
             entry
            .     .
    predicated    allones
            .     .
             exit

if the basic block is in itself a loop of a single block, then
we handle it differently, separating it into two loops, one that is executing while the mask is allones (full), and another for the reminder.

The allones optimization takes place in the predicator, but several other changes were needed in the vectorizer as a result.
1) vectorizer heuristics: was edited in order to simulate the
same results as without the allones optimization. That is, if the heuristics chose to vectorize (not vectorize) a kernel before this commit, then it will make the same decision after this commit.
2) WIAnalysis :
was updated to analyse allones branches as uniform.
this is the correct way, because all work-items will choose the same branch. This is also necessary to maintain high performance otherwise the packetizer works too conservatively and loses some optimizations.
3) packetizer:
the packetizer misses some optimization opportunities as a result of the allones bypasses. This was fixed. Some missed opportunities of the same types also exist without the allones bypasses, but in this commit we only fix the ones caused by the allones bypasses.
These opportunities includes:
a. LoadTranspose opportunities when the user of an ExtractElement is a phi-node.
b. StoreTranspose opportunities when two different store instructions exist for the same value.
c. (this is not a missed opportunity but an inefficient thing the
packetizer sometimes does):
constructing a vector to be used in a phi-node only to later be dismantled back into scalars.

Many performance implications are expected, mostly speedups, but some degradations are inevitable. Also, a slight increase in compile time is expected at some cases.

Review: (Amjad)
http://cvcc-ubu-01.iil.intel.com/cr/r/4143/

Precommit:
http://compute-ci-jenkins.iil.intel.com:8080/job/PCompilerReport/310/