Remove OpenCV references from JumpStart samples
Description:
Replace the OpenCL usage with proprietary solution for read and write image (bmp) files.
This come also to fix CSSD100019244 - Can't run JumpStart samples in VS2010 debug mode.
Now, the OpenCV is not part of the SDK installation and there is no reference to it in the JumpStart samples. Instead, a very simple bmp class is used to read and write bmp files.

Review: Ofir and Simeon
http://cvcc-ubu-01.iil.intel.com/cr/r/4352/

Performance Implications: None

Testing: Manual tests - generate, build, and execute all the samples in VS2010.
http://compute-ci-jenkins.iil.intel.com:8080/job/PToolsReport/215/

CQ: CSSD100019244