the spir code generated for this test contains global data that is initialized with an absolute address of another data in the jit buffer. this requires load time relocation. this patch takes the relocations generated by the code gen (http://cvcc-ubu-01.iil.intel.com/cr/r/4701/) serialize them on the host, deserialize on the device and patch the jit code accordingly. note that the use of these relocations is different than the use of relocations to external functions, therefore they are held in a separate vector.
move to llvm 8433fa2e_654f5c67.

in addition there was a bug in the spir test that creates a read write
buffer is read only. it works for the tests on the CPU but doesn't
return the data properly if the test runs on MIC.

CR: http://cvcc-ubu-01.iil.intel.com/cr/r/4702/ by Mohammed
llvm post commit:
http://compute-ci-jenkins.iil.intel.com:8080/job/LLVMPostCommitReport/107/
pre-commit:
http://compute-ci-jenkins.iil.intel.com:8080/job/PCompilerReport/654/