Fixing ticket CSSD100019557. (recommitting and fixing a bug of previous commit)
This is a functional bug in spec-accel/heartwall.
It is caused because of a difference in what the WIAnalysis outputs
when running before the predicator and when running before the packetizer. This fix makes the predicator
more WI-friendly, which has the potential to help performance in addition to fix the error.

The changes are as follows:
1) When converting phi-nodes into selects, we locate the condition of the phi-node to be used as the select predicate, instead of using the mask. This is important because the mask is always random, while the original condition could be uniform (if it is a uniform condition inside a divergent block).
2) Before this change, users of an instruction inside a loop uses a special value created inside the loop that saves the values of the last iteration of the loop. Now, we only switch the used value for random user. If the value is not random, it means all the loop exits are uniform (even that the loop itself is nested inside a divergent region), which means we can safely use the original value.
This is a conservative condition: it is possible that all the loop exits are uniform and the user is still random. In such a case switching the used value is unnecessary, but not as harmful.

Pre-Commit:
http://compute-ci-jenkins.iil.intel.com:8080/job/PCompilerReport/715/

Code-Review:
http://cvcc-ubu-01.iil.intel.com/cr/r/4740/

Performance implications:
Some speedups are possible, but not expecting anything major.