Removing the mask off consecutive loads of local memory in kernels without local memory input parameters.
In general, since unmasked loads are more efficient than masked loads, it is tempting to do every load without a mask.
The problem is that this can cause exception if loading from a memory without permissions.

In this patch, we enable to do unmasked loads instead of masked loads, when the following conditions hold:
(1) The loads are in local memory.
(2) The function receives no arguments that are pointers to local memory.
(3) The addresses are consecutive.
(4) The block is being zero-bypassed.

(1)+(2) guarantee that the load is from a static buffer on the heap. This patch also adds padding
to this static buffer.
(4) guarantee that if this instruction is executed, the legitimate mask is not empty, thus at least one address is ok to load.
(3) guarantee that the rest of the addresses are 'pretty close' to the legal one.

code review:
http://cvcc-ubu-01.iil.intel.com/cr/r/4874/

pre-commit:
http://compute-ci-jenkins.iil.intel.com:8080/job/PCompilerReport/803/

Performance implications:
Some speedups are expected, specifically srad benchmark of spec-accelerator. CPU is expetced to be more influenced than MIC.
No regressions are expected.