[Builtins] Fixed compiler issue with non aligned constants.
The constants were allocated as arrays of scalars without any additional requirements on the alignment.
Built-in implementation casted the pointer to the constant array memory to the vector data type pointer, which must be aligned by the size of vector by specification.
By default clang aligned those constant arrays to 8 bytes on Windows, which is not enough for float4/double2 data type and leads to crash on CPU with SSE4.2 ISA.

CR: http://cvcc-ubu-01.iil.intel.com/cr/r/5027/
Tests: compiler pre-commit.